feat: Credential Management - Credential Status API
