+++ api/spec/openapi.gen.go
@@ -19,86 +19,89 @@ import (
 // Base64 encoded, gzipped, json marshaled Swagger object
 var swaggerSpec = []string{
 
-	"H4sIAAAAAAAC/+xce3PbOJL/KijeVc2kSpY8yeT2VvfPeiTvju6S2Gs73rrapFQQ2ZIwpgAOAErWpvzd",
-	"rxoPEhRBPZJ4LlO7/9ki8Wr049eNH/gpScWqEBy4VsnwU6LSJayo+fMiTUGpO/EA/AZUIbgC/DkDlUpW",
-	"aCZ4MkzeigxyMheS2NeJeZ/4Bv2klxRSFCA1A9MrNa9NNb7W7u5uCcS+QcwbhClVQkZmW6LxUamXQrJ/",
-	"UHydKJBrkDiE3haQDBOlJeOL5KmXpFMueBqZ7615haSCa8o4/kmJeZVoQWZASgUZ/plKoBoIJYUUYk7E",
-	"nBRCKVAKBxZz8gBbsqIaJKM52SyBEwm/lqC07TKVkAHXjOb7pjeFx4JJUFMWEcWEa1iAJBlwYXpFAeRs",
-	"DpqtgDBcfip4pnA2+Mj1GYzHbA844L6B7vb3G25HvHMJcwlquW9P3Su2lx7ZLFm6JCnlocjFDLeEcNg0",
-	"xlRRCapUFJHtvbq+m1y9u3jTI2xOmNmClObYOy7FNPIbVWtVmjPg+r+I0EuQG6agR24u//p+cnM5jo5t",
-	"pjW1P8cWi0+89EItjnRmpPdrySRkyfDvTeNoDPSxl2imc2wbs8uqYzH7BVKd9JLHM00XCjsVLEt/XKfJ",
-	"x6deMqr08lZTXaqrQkeEaP5Qxqpx5sZgGgrdNGkvif1r211GbCr7lmFkKM0qxiwbCT5ni/bcx5Mxsc+I",
-	"7PRBf0Lzh8fI0t2D6LbnjD9ANs1YptoNryUo4No6JsbJLxv1vW36gghJflGC59n3dlkvULIrqlXSS5iG",
-	"lelOcLiaJ8O/t8TY+7QjlCcUgfuJSkm3RtRerJVsjtKNDNa0YEaoPwPN9XK0hPSh2937J2RV+f2laUdS",
-	"bEiU2cq2xNNSSuD6jq0inY7sQ2IckLOb2rdbWSXDJKMazvCdqD8wA8f8Pf5OmCIfElUay/mQoHOwA+CD",
-	"siCUZ0SWHOPBYSN1QwWqHBNdTOpKiyJni6VRPJYlw+QPv5TqMV+l8tUPi9e4jnprrFyNWM3+TDjTjGq4",
-	"moxHP96PbqwX2xeRfQuCTUhtb2SiVEkx5rlOIlE6DLTTDDRluYrtnNJixf4BimyWVJMHxjPcQee/J8Zk",
-	"yYZyjc6cLNjaxNn70W08LOaUrabAs0IwHlnaCJ8T/9yrihtlLsUKI7GEwFkR0yXJqKZkSZUL8nUEoHMN",
-	"kjjFmJd5viU0xa02YfxgFLKRY8qcoKfMCXZayrw9/fc3b/yc/YvENUWvEa6Lkr/RPAfdJ3f0ARQpJKS4",
-	"phSIQLV1A28gzx+42FTogxRU0hVokH0ymZOZQMvcM0mj+a3OqATChUbss2YZRmcLcJzt+57qVeDKNizP",
-	"Pa4iqVGMjjcZdw6QiAI4y878a2f+teFgsE/e1UyPwXcbI8jBUuQZSEKLImepFbgxC9slqRefGu9ZSvvO",
-	"+5s38ZlUKjbVsCpyI9gsggfcwwql1appddEB3M2S5dBUxFTwNC8zC/2YIoi4JE2x434FUAzQwY4LKebY",
-	"BVPVCiysKtFbl7lmRd4c3s0srtkLSbnuwDjO4BC+OQ3x+21aGfyjiF5KUS6Wdu6BWt7h//WLgVkaIGgF",
-	"AY/pkvKFsULezAjQszXzgFRkBrjiaiRRGgpltL+twhnMaZlrHK/p4bCLqBxEMUV/351JrGlegksXKkS5",
-	"42tR79AxFvTXEjwYtQZONPpNjEIOBc/QhZoIWM7OFFo112ayFsuaBXtj3zC97BgPV0gcmiEKNEa5rDQz",
-	"LiSsmShVIKkaBRN0NGwNilC3NJR3cw97hGny9v3tHWFGQwH/Z9zP2k/6ojlpF2v88iMiUuaBl3g9np1I",
-	"3w757uqu0hXGzSC1JoxQE+a52NiUrJBw5vcZsqnVE+NMEatF99s7uQ7VH1m/ompnaHTYbaJZBjwWkGqF",
-	"Qc6bn9XpAiS6PdwC43maSuz2tE/GVkeNUewmXQfzn2p+5rk6bmJh5to2LNz/Ooo252f9dz/Ese1sqQup",
-	"dgCaI3OAVuvDFYpj8FBXynBkiDezGU06Al/g5pyjqK21oAq1Koc1ekbGbYTEXdjxFyLSOap8n9yWRSGk",
-	"Vjbs/3x3d03+cnlnXI/55wYyJiHVfTesIiu69dZA/npj9zsInd7PGPiEEiwVttKCKHT+BnHpJTBJVmKG",
-	"muTmSIsinjg/xmNkQyzeG9SB2maiqZAScpdgzQkHyI5JqeMb5+fycY86HpE+derj9ZhqiittalERZInT",
-	"DOZmcoI7qbTkVZSyEKojt47P2w68O+f2/EJ72Qf9ZZ1mdG/ngaJGtMODO3DduQO4GhR8bcFe2l2W7zzg",
-	"viJG/SziVWtXcWxqjwOcmNILW3bB4f9dwjwZJv82qIuzA1eZHews3RVrWlIP1hOKOCK3YxU8Pu4XV4/S",
-	"JToNvohFsyXNKV+YoE2zzAIkB3bFvAuXIxCLF+ayAIjbLhD8iBXTiN3UVmlY2SqESWacUzqA/2+rysO+",
-	"XYtVu556SSZWNFaTHZvfT1j3GiSbO3/5FvRSdIjg/c3ES6DdxPpgC/hiEpozqTSB7OXr1z/8kRTlLGep",
-	"qYSLORlPxuR757uFJNcuHRlPxi8OSfOpUz+9kh2potelWkLWgITH4ALbbAdKdkOBfbX0C/Lft1fvCC9X",
-	"MwzriL4kOK+vmhV8tws+xuLGBMV3igC8EIpptgbiiu2YPDVb1HV7Rag2HWZMpRK0O6foOjUhs1LbfdHb",
-	"gqUUE05TGcEIu4Z8S9RSSE2+h/6i3yMz0BsATl4bNPEf5+d+oi+6jgTMHKelZF0HAvUiTGRHadtcV0Qm",
-	"XZU3hNKQuSTSiAzlpBhf5HCGiFbCHCS48xwrX4WwF6XYgDPtfCWOxw8CjHCpjYOWwO3uU8xjS/cOH1/Z",
-	"tz6v8G0SwmhlzT5AjamLY9WEWt4mhuLYAdNuzv/IovQ9+qht7RB8jbXMI4uwL4c1jipLC0DtnLK8lOAK",
-	"1q60GQtLkD7EQhK2MmuK+mGQUsh2s0v8maxAKbqAz3bg98E7ZGVeOqyfdiF+ZtGBAk3dJ/B9e2Z7dX54",
-	"t5ND8CzYsXB2v2OQtiuB01BaVH6fLf2jkNp613aeG6h9JeTz1C21Y8DDXsEdgx0qDyPmgfDUIT1Gq7J1",
-	"BV83OUWbQqPcV2HpXNCJIgkPN4/xwGGa+/vxwXv9Zss6u2TyBaI95CYbYt2vYCe5qXAOlaPqNYoVX+nA",
-	"+2SHu7MnjSnt3ZLPcZkxORzjNMNZnew2zaNvwG/GFv8F8jvVd56g25/lPLvM9bD7jK7qSMlgb4zPhcfe",
-	"1CJ3WFGWJ8NkCXku/qRlqfQsF2k/g3XSSzhdYc93+PNPuUiJBrpCMZiKc7LUulDDwaDZDDWjmWRVze9H",
-	"t0TZ0nCItKriMWZ0ocRJifkU+durEbkfnV1cTwjNBV/Y06arAvhk/OP9CDVLi1SEBbaB6QZkeFRpm7nT",
-	"7KSX5CwFpxduoRcFTZdw9rJ/3lrjZrPpU/O4L+Ri4NqqwZvJ6PLd7SW26etHu4vhpjE6yxtHTLcg1ywF",
-	"8v396PaFBcHKyum8jwMbZAcc045h8qp/buZSUL006jUIuRjDT8kCdIwao0vJlc+vO2gxqMhGypMsGSZ/",
-	"Af1z0HV9/GSGfXl+vpO0BUXxAXrYmrF5yApiFBWjnjve7X+MBahytaJyW1FbyMjNL05OeeolA6cB7hha",
-	"DT65vybjp0EAjOx7pursj/eUCQbRI2ZfRZoYIh66PbMptZFUgyRhoNCyhF4gmF3f97GXFCLGnZns1okj",
-	"e3YtlN4pU6mkqnj8JLLtV9u0WLX26enp6Qv1ZCfKHqEFZiKBWAJNqMpvh5Qg8AlqUDE+BMvSb1Ub3gcH",
-	"daw6qGPhKeIk6lFD98e4oTu5yl0TCqsuKk2MHlu9itMxZ9zm/I21j8v26Kx0MlTXXlSTYFuqAxjclGdS",
-	"6fih71fQ6pNGPcEDHjw17jIGVO1BRT/ojBtXF6Vekpf981YB2ojGk6XdZlgenWHPVKyYXT6FJ5JF4w3u",
-	"7UU1qd5+w7s3DBDDu5iBoZFoQT4kqcjgQ1IZ4a8lyG1thU0KxSmW2ItVih0vxZK5EVd1jet5btmXjXlB",
-	"KsROMpBsDVl10m1PyD3Aq7g75vzcFZSjVeSeO/p3LTNCF2j+2lKWOhckMpjW6cMXrsrWC+2cN7QmHNk1",
-	"urN/P9hxU5raPpOT9zR6IiEdPcGek5QK5BldGF6wCAhK36nqxQZp0rOa8i0BpeksZ+ZQp6JIRYd0jKgG",
-	"/WnBlHbkv0IKY2JCWj7Rij741zsPC+IWYSfszghOFJa9tdC8jXFgQMsOOk1BuOenWdpXyFJxstGCYI5p",
-	"CJ6WouWPhcKDLMMopXk+o+mDDXtR0TvqmLLcMjumu0fjdtdJOlAE7LKpDXaAmil2+/PV+zfjKmy6qska",
-	"XYc5hhZKnSmm69nOhVyA3HYK0nD/vky/PX0ao/4atla9/W90Jkq9wwy1b1jeak2ettdj+uSt53J2DBKg",
-	"Bqv85jaKIWpNm+zTasca+8M4SalNyiO0UdUlqThj/CTJ2STxO+WSTDISnEOqPSPo/c0bu92eVs7y3LDY",
-	"/KmnWIPcVkZrXJsGuWIcAoF+hyIq6IzlTDNQRl29E1F9cnM5unr79vLd+HKMkhhvOV2xNIyuN/tNz44y",
-	"dUjgM00QdZ4s0a0FmvD24n/NctH66lNLb2qOj6fZiv0DKsP5Thnan2TAU/gKq8M+pzix01bW4Mq6SL51",
-	"N+VAGofits2Tt+FR+wPtHYQNsk8uOrmpGI7rE+2CKscTpTzKua/cgA/wNc6vJe+Om1sU+5C2a+h72KTm",
-	"r9opNnxWeyV39ZirUmmi6YPJHwR6elFyRxCuOmXK0JoXJUUMCO7mnGQLxvGxWwdTrtMeSUWZZ+gRKCdU",
-	"a3TKHXtbUZ0PJEsNlP7q/OUelP54ttlszuZCrs5KmQNH+JA1YfvuAXUWu3Xkb+NFwonBLQvgCHMPXBTt",
-	"am3wrT24t6yHfOsuhjAD7xxFH8Mf02zhczDJ1AN6yRzoQ8clxThz3C/HU/c/2Bc/JIFqIULzzEyHLF0U",
-	"7qAN49rgkaba6Z3jcYfY1UbMwwcfuAcfe4fLBH8WJc92sqVoDhNkSDWjoUqRCmoOa+I5+MguWgHPFLGp",
-	"UpwZYkFDvm1x+z3gQEe/AK12GTc1kxsNLQyfVLXpJJ47EnhgWRPbmwN352GYj2Midk3l0Xn2ybZ05B2u",
-	"fwK80nlvq4MkF005250007Pht5FIHpimT9mGXyFB/NzLMv8CAP//ACBy8SXIUYf/ZEn7qdeADtbFjr2v",
-	"c0R+fyyq+FcC35JUna0Mv/FcqzX1Zho5/N2nyoc4tM2ScVjK3QmzMWDaPjr44asdHeyj7kYw8cjdP3jq",
-	"JT+ev46wH2yQfSc0uchzsXGv/vAqdqBsNfySa6a35E4I8obKBZgGL/8Yu0clyFvKt17uagebd/Dcj4Do",
-	"1Zdd9hybquY1XXvhFV8wgZg2PwFTXTJwF32N/4ukZhaTW++1pMqBysjF/ZZvrynQceh9576x8kzgew/C",
-	"bF3d9nDTRvbwrq9nXZz27aNoDt2R+7bTxJPGmVZklxj8ldtCi4WkxdLBMUl5JlbE9tG6+uxv+MGeCwAu",
-	"eFgl2hcl911g7wjf7SvhHcF8P1Zq7e+HRoNWmu8CaXYA+FN7L5y569Cqmn/1maPDfjYQitOT3W08zr1+",
-	"vZPZ2MeMOk5l0aOet8X9E81IfYwc+jv/KbL9Ds4J+Myud/CpLFn2dJDl4/XRtmq7mcZth5+270t3LPlM",
-	"YmxerjjmWNtHl+ZCEDCWdq6R2xh7z4qxWXB/pL49EiFqlOWJp7Q4+sAraZNK0jCXMxnea3XBqrkzhsrU",
-	"dRXnmaIBy7o+jjYZuy/mGUdr733x9vceUmCFReYV7F6BpqYaVZdk7q9tZ6ekDbe6ShvjHnfn7nX07krR",
-	"tTw/o3raggOmZishgQRcvZBlqTr4qkf6pZ31lehgcJavY4//bJnoO8bhmT9VgaZxG71BArL4u0FfNeQc",
-	"//mT+9FtYEwhNzTU6IO0OUs1PsyUGlsyrenjmUhSbTy1e8/gufhw0XsxzxygOu9QHONkWzfSPlcVvoQ8",
-	"95urREgjC+hlvwmj7Pq30ImuDxUcoxLfhmsJO/1dOJcwPjyrd2ldIvlN/Ev0ksEJHqZoiqdDJ7wG3G0L",
-	"eIorxgby/Mx8CG2QsewsrT6zuRcN16+2kXD9sc5nlGI9yCnYN1jh6XjXs5HvbDbVbQx3X0yDrIjP2Ve1",
-	"OiMUk9/b9dV3L4aDQS5Smi+F0sP/PP/DeYI26iTUuvhtaoxnNnXP7JdBd4qF9VRdQbO9Rq+qR/ZTaXa7",
-	"p8gFjLpdeHHh6ePT/wUAAP//J98z/v9aAAA=",
+	"H4sIAAAAAAAC/+xcbXPbOJL+KyjeVU1SJUueZHJ7q/uyHsm7o7sk9tqKt642KRVMtiSMKYIDgJI1Kf/3",
+	"rcYLCYogJcV2JlM7n+KIeG00up8GnsbnKOarnGeQKRkNP0cyXsKK6j/P4hiknPI7yK5A5jyTgD8nIGPB",
+	"csV4Fg2jdzyBlMy5IKY40eWJq9CPelEueA5CMdCtUl1sprBYs7npEogpQXQJwqQsICG3W6LwU6GWXLBf",
+	"KRYnEsQaBHahtjlEw0gqwbJF9NCL4lnGszgw3mtdhMQ8U5Rl+CcluihRnNwCKSQk+GcsgCoglOSC8znh",
+	"c5JzKUFK7JjPyR1syYoqEIymZLOEjAj4pQCpTJOxgAQyxWjaNbwZ3OdMgJyxgCgmmYIFCJJAxnWrKICU",
+	"zUGxFRCG0495lkgcDX6ybXr9MdMCdtjV0bS7XX85wo0LmAuQy641tUVMKz2yWbJ4SWKa+SLnt7gkJINN",
+	"rU8ZlKCMeR5Y3ovL6eTi/dnbHmFzwvQSxDTF1nEqupJbqEqr4pRBpv6HcLUEsWESeuTq/O8fJlfn42Df",
+	"elgz83NosvjFSc/X4kBjWnq/FExAEg3/Wd8ctY4+9SLFVIp1Q/uybJjf/gyxinrR/YmiC4mNcpbEP6zj",
+	"6NNDLxqVenmtqCpkcwJVCSJ1keYWlmXVpmisULqnaRuwxb2pNYbXNS8tVBGc1kWuArqh/5DaWGFdbQdq",
+	"+7Q+zcPmsm8KOJQDZzFmyYhnc7Zojn08GRPzjYhW0/oXtGpwH5i6/RDU5pRld5DMEpYEtOFSgIRMGXvL",
+	"MvLzRr4wVV8SLsjPkmdp8sJM6yVKdkUVrhpTsNLN8Qwu5tHwn01d+bwjlIdPpf5EVAi61aJ2Yi1lc5DK",
+	"J7CmOdNC/QloqpajJcR37V7MfSGr0p0tdT0SY8XWnRAXQkCmpmwVaHRkPhJtV605qFyWkVU0jBKq4ATL",
+	"BM1cyy41ukWYJB8jWWiD8DFCm2c6wA9FTmiWEFFk6Ob22x7blafKIdGFpC4Vz1O2WGrFY0k0jP70cyHv",
+	"01UsXn+/eIPzqJbGyFWLVa/PJGOKUQUXk/Hoh5vRlTHOXUDD1SBYhXj2aiJlQdGV20YC4MPHD7MEFGVp",
+	"yAYWUvEV+xUk2SypIncsS3AFrVua6C1LNjRT6KPIgq01fLgZXYe9fUrZagZZknOWBaY2wu/EfXeqYnuZ",
+	"C75CgCHAM1ZEN0kSqihZUmmxS+XY6FyBIFYx5kWabgmNcak1OtnrXI1DnDEr6Bmzgp0VIm0O/8PVWzdm",
+	"V5DYqmg1/HlR8g+apqD6ZErvQJJcQIxzioFwVFvb8QbS9C7jmxJUkZwKugIFok8mc3LLcWd2DFJrfqMx",
+	"KoBkXCGkW7MEQYfBbXbvu5aqWeDMNixNHVwksVaMlpIsswaQ8Bwylpy4Yieu2HAw6JJ3OdJDYOtGC3Kw",
+	"5GkCgtA8T1lsBK63hWmSVJOPtfUshCnz4epteCSlis0UrPJUCzYJwBz7sQSflWoaXbS4fbNkKdQVMeZZ",
+	"nBaJQbRMEgSSgsbYcL/EXRq/YcO54HNsgslyBgYtFmiti1SxPK13b0cW1uyFoJlqgW52wyEqtRri1lvX",
+	"0rBOErUUvFgszdg9tZzi/6uC3rbU+NYIAu7jJc0Wehdm9UAHLVs9vIl5ovE4zkYQqSCXWvubKpzAnBap",
+	"wv7qFg6bCMqB5zO09+0B0pqmBdgoqATKO7YW9Q4NY05/KcBhbLPBiUK7iV7IgvtbNKHaAxa3JxJ3dab0",
+	"YA1E1xN2m33D1LKlP5whsWiGSFDo5ZJCjzgXsGa8kJ6kKnBP0NCwNUhC7dRQ3vU17BGmyLsP11PCtIYC",
+	"/p9lbtRu0Gf1QVtf46YfEJHUH5zEq/7MQPqmy/cX01JXWKY7qTRhhJowT/nGRJq5gBO3zpDMjJ5oY4pY",
+	"Lbjezsi1qP7I2BVZGUOtw3YR9TTgPodYSXRybvsZnc5BoNnDJdCWp67Edk37ZGx0VG+K3Vhyb1hXjk9/",
+	"l4cNzA/ImxsL17/yovXxGfvd93FsS6QTQKotgObAGKBRe//ByyF4qC1kONDF69GMJi2OzzNz1lBUuzWn",
+	"ErUqhTVaRpYZD4mrsGMveKBxVPk+uS7ynAsljdv/aTq9JH87n2rTo/9zBQkTEKu+7VaSFd263UD+fmXW",
+	"23Odzs5o+IQSLCTWUpxINP4acaklMEFW/BY1yY6R5nn4POA+7CNrYnHWoHLUJhKNuRCQ2gBrTjKA5JCT",
+	"gvDCubF86lDHA8KnVn28HFNFcaZ1Lcq9KHGWwFwPjmdWKg155YXIuWyJrcPjNh3vjrk5Pn+/dEF/UYUZ",
+	"7cu556wm2ODeFbhsXQGcDQq+2sFO2m0731rArkOM6lvnUc+hoT12cGRIz82xC3b/nwLm0TD6j0F15jyw",
+	"B86Dnanbw5qG1L35+CIOyO1QBQ/3++jTo3iJRiNbhLzZkqY0W2inTZPEACQLdvm8DZcjEAufNyYeEDdN",
+	"IPjhK6YQu8mtVLAypxA6mLFGaQ/+r84Hu1YtdNr10IsSvqKho+ax/v2Iea9BsLm1l+9ALXmLCD5cTZwE",
+	"mlWMDTaALyShORNSEUhevXnz/Z9JXtymLNYH/HxOxpMxeWFtNxfk0oYj48n45T5pPrTqp1OyA1X0spBL",
+	"SGqQ8BBcYKrtQMl2KNB1RXBG/vf64j3JitUtunVEXwKs1Zf1iwm7Cs7H4sJ4dwoUAXjOJVNsDcTeIWDw",
+	"VK9RXUdIQpVuMGEyFqDs9UvbZRC5LZRZF7XNWUwx4NQnI+hh15BuiVxyocgL6C/6PXILagOQkTcaTfzX",
+	"6akb6Mu2mw49xlkhWNs9RzUJ7dlR2ibW5YFBl8cbXCpIbBCpRYZykixbpHCCiFbAHATYayojX4mwF6VY",
+	"gzPNeCWMx/cCDH+qtfsjz+x2KeahNxIWH1+YUl928K0DwuDJmvmAGlMdjpUDalibEIpje7Z2ffwHHkp/",
+	"yNFm7xrP1rPPzuLl+kslilgZQIsVUOtuRu1n12Vzk3EQBz3WF3T47sk4CrTvaVa3gA60mzfoCbZVI+4k",
+	"u0gDMjaF/ZOkMhb2Qoc5ZWkhwF4L2APkkPOH+C7k+LGWnmPQ24EQXDSrnePPZAVS0gV8sZu88cqQlS60",
+	"3wqYibiRBTvyVq1L4F1rZlptWbV9INhbMX90v2MovCuB47BwUH5fLP2D8PB6d+88Nxx+Inz50C61QyBa",
+	"p+AOQWilheFzT3hynx7jrjKnN+506hht8jdl1zlW64SOFIl/hXyIBfYPE34/NrjTbjZ2Z5tMHiHafWay",
+	"JtZuBTvKTPljKA1Vr3Yk9ES0gqMN7s6a1IbUuSRfYjJDcjjEaPqjOtps6k/fgN0MTf4R8jvWdh6h219k",
+	"PNu2637zGZzVgZLB1lg25y7CoSY+ghVlaTSMlpCm/C9KFFLdpjzuJ7COelFGV9jyFH/+MeUxUUBXKAZ9",
+	"rh8tlcrlcDCoV0PNqIeyZfWb0TWR5gDeR1rlET3Gzb7ESYFRK/nH6xG5GZ2cXU4ITXm2MHd6Fzlkk/EP",
+	"NyPULMVj7h9jDnQzIPwLYVPNcgaiXpSyGKxe2Ime5TRewsmr/mljjpvNpk/15z4Xi4GtKwdvJ6Pz99fn",
+	"WKev7s0q+ovG6G1au8i7BrFmMZAXN6PrlwYESyOn0z52rJEdZBjcDaPX/VM9lpyqpVavgc94GX6OFqBC",
+	"BCRViEy6U4wW8hEqspbyJImG0d9A/eQ1XV3y6W5fnZ7uhMbe1cMALWxF9923C0JEIK2eO9bt//QOkMVq",
+	"RcW2JBCRkR1fmAL00IsGVgPsZb8cfLZ/TcYPAw8YmXL6bN9dokrtDIIX+e6sbqJZnGj29KJUm6TsJPId",
+	"hRIF9DzB7Nq+T70o56EofbJ7Gh9Ys0su1c5hoIzKc6UfebJ9skULnYk/PDw8PFJPdrzsAVqgB+KJxdOE",
+	"Mlg/RgkqTty3qAXm5EL6xrJtA6MyeHpQUl6fQxu6D1R+E73olNQTaMjgs/l3Mn7osruCwRrkLoupw+iG",
+	"luw31MRemBuqWwl0IquvR2n7V1aOAxbmaBXxgIUclOQ8zpL4mzUmHqeClZwK5hM+JkFY5mMolmlmqr1k",
+	"qcfTso31GErQKIvicDQdSVMlWJPZ0OH4hJWhvHSimnjLUt6V46I8k18M83OewAQe1esRMGovwadtM6Bq",
+	"D0qmWKsRvDgr1JK86p827gq1aFy6jl0MQ3nWRMeSwLhLfXOc36D9xLU9Kwe1x3beaLKepsjdgmb8KU4+",
+	"RjFP4GNUbsJfChDbahfW2W6PMqbTikJo0okwOGvr11GSk8f1eUbKsJ8kINgakpKUZMhMLkosaZaa6mTv",
+	"/oIXfj3L0rI1E0IXuP2VYZe2TognMKvOIB45K3PpYMa8oRU31MzR0rRcZ4cNaWbajI5e0+DlsbBMMuN0",
+	"CgnihC50Cgf3uKTfybJgjd/uCKjploBU9DZl+v69ZLMGu7Tk1RpTdcGksjztXHC9xbgw1M8VvXPFW+91",
+	"wzvCDNhe5x4pLJM3V88H3NOhIXIepyCZoxIbhq5PKLSyUZysKDNcfMOmdTf4PudAk/9pmt7S+M64vaDo",
+	"LctXGhqw6dNmctrVtZL2FAGbrGuD6aAi9V7/dPHh7bh0m/bodY2mQzOGuJQnkqlqtHMuFiC2rYLUNO3H",
+	"6bfLdEGvv4atUW/3G73lhdpBWaaESTGo8lxMgmafvHO0+5ZOPNRglF/nQ2pO7ayeKFCuWG19WEZiak72",
+	"Agx/2SapcHLPUZIzJ03fSXtSRUY8yyBWjrz54eqtWW6XAcTSVBOOHUGFr0Fsy02rTZsCsWIZeAL9DkWU",
+	"01uWMsVAanV1RkT2ydX56OLdu/P34/MxSmK8zeiKxb53vereeqaXmUUCX7gFUefJEs2apwnvzv5fTxd3",
+	"X0UwcVvNUqcVW7Ffodw430nN0BYMshieYHbY5gwHFh0ZFHlpDdaTb22uNghtUOyyuTwbuFeOe7SDsEH0",
+	"yVlrGgG644p8lFNpKf00C6ZHlWbAOfgK51eSt8ygRjaUn2GhmdZYpUo1MEOs2azmTKZVn6tCKqLonY4f",
+	"OFp6XmQ2l6NslEmdgbIoKGJAsLnbgi1Yhp/tPJi0jfZIzIs0QYtAM0KVQqPcsrZlVsoxsejr01cdKP3+",
+	"ZLPZnMy5WJ0UIoUM4UNSh+27XKIklCDq8sED7kTjlgVkCHP3PFXQVlvjW8OxMgS1dGtz+JiGdzabCt0f",
+	"U2zhYjDB5B1ayRToXUuafDjJx03HZVl9NAU/Rp5qIUJzJHqLLK0XbsnwwLnBPY2V1TubcuNjV+Mx99+e",
+	"4hp86u0/NvgrL7JkJ1oKxjBehFSRz8oQKaf6xjccg4/MpCVkiSQmVAqT+AxoSLeNNCwHONDQL0DJXXJk",
+	"lXSDG813n1Q2mX+O5udZYFHlINU7bo/DMB7HQOySioPj7KP30oHptv8GeKU1xbaFzxwMOZuN1MOz4bcR",
+	"SO4ZpgvZhk8QIH5pXuMfAOC3BwCBHEUvRh3+mwXtx2Zs7j0XOzS18oD4/lBU8UcA35BUFa0Mv/FYqzH0",
+	"ehg5/N2HyvvSHepHxv5R7o6bDQHT5tXB9092ddCVZRHAxCObKvbQi344fROgUBkn+54rcpamfGOLfv86",
+	"dDtqNPw8U0xtyZRz8paKBegKr/4cSnnl5B3Ntk7ucgebt6QkHQDRy7fFOrgXsv6ignmbAAtoR0zrj5CV",
+	"+WD2TQZt/wKhmcHkxnotqbSgMvDGSsO2V9kqYeg9ta98PRP47kCYjVc2HNw0nt1/lsFRt457fS8YQ7fE",
+	"vs0w8ah+ZiVjLgR/xTZXfCFovrRwTNAs4Sti2mi8UuGSsaEjV8s6D6NEXV6y662RFvfdfL2jxZl3Y6XG",
+	"+n6sVWiE+daRJnuAPzVPeDD7coUsx18+tLffznpCsXqyu4yHmdenu5kNPafXciuLFvW0Ke4faUKqa2Tf",
+	"3rnHMLsNnBXwiZnv4HNRsORhL1XQ6aOp1TQztcS0H7cfCnst+UxirOfBHUbxqOevmYkgYCzMWAOJc513",
+	"xVjNS/WrEv0CRI2iOPKWFnsfOCWtU0lq2+VE+E8QWGdVXxnNh2zLmnwmb8CStuc5J2P7Zqs2tCZFN2s+",
+	"zRMDyw0yL2H3ChTVp1HVkczNpWnsmLDhWpVhY9ji7jyTEUyAy9um50ZUDZtngKHZigsgHuHXp2rLFtL7",
+	"gXZpZ34FGhgc5ZvQ57+adJZdcpyNncoDmtrDITUSkMHfNQ68Jue4l6puRtfeZvIJ5r5G7yXVmXyF/Uyp",
+	"sWHk6zaeiSTVxFO7yUrPRaMMJtc9s4NqTcQ6xMg20lq/VBUeQ5776irh08g8etlXYZRdfg2daHtT5iDe",
+	"7TdhWvxGfxfGxfcPz2pdGploX8W+BDOVjrAweV08LTrhNGC6zeEhrBgbSNMT/WblIGHJSVy+iNyJhqui",
+	"TSRcvav8jFKsOjkG+3ozPB7vOjby1ERT7Zth+mgaZEl8Tp5012mh6PjezK9K4BoOBimPabrkUg3/+/RP",
+	"pxHuUSuhxhsd+ozxxITuiXnEeeewsBqqPdBsztGp6oHtlJrdbCmQxVXV87OfHj49/CsAAP//fGwDtYFh",
+	"AAA=",
 }
 
 // GetSwagger returns the content of the embedded swagger specification file
