--- pkg/doc/vc/status/csl/testdata/examples-v1.jsonld
@@ -1,16 +1,0 @@
-{
-    "@context": {
-      "@version": 1.1,
-
-      "id": "@id",
-      "type": "@type",
-
-      "ex": "https://example.org/examples#",
-
-      "image": {"@id": "http://schema.org/image", "@type": "@id"},
-
-      "CredentialStatusList2017": "ex:CredentialStatusList2017",
-      "DocumentVerification": "ex:DocumentVerification",
-      "SupportingActivity": "ex:SupportingActivity"
-    }
-}
