+++ pkg/restapi/v1/issuer/openapi.gen.go
@@ -11,6 +11,12 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
+// Credential status.
+type CredentialStatus struct {
+	Status string `json:"status"`
+	Type   string `json:"type"`
+}
+
 // Options for issuing credential.
 type CredentialStatusOpt struct {
 	Type string `json:"type"`
@@ -82,14 +88,28 @@ type IssueCredentialOptions struct {
 	VerificationMethod *string `json:"verificationMethod,omitempty"`
 }
 
+// UpdateCredentialStatusRequest request struct for updating VC status.
+type UpdateCredentialStatusRequest struct {
+	CredentialID string `json:"credentialID"`
+
+	// Credential status.
+	CredentialStatus CredentialStatus `json:"credentialStatus"`
+}
+
 // PostIssueCredentialsJSONBody defines parameters for PostIssueCredentials.
 type PostIssueCredentialsJSONBody = IssueCredentialData
+
+// PostCredentialsStatusJSONBody defines parameters for PostCredentialsStatus.
+type PostCredentialsStatusJSONBody = UpdateCredentialStatusRequest
 
 // PostIssuerProfilesProfileIDInteractionsInitiateOidcJSONBody defines parameters for PostIssuerProfilesProfileIDInteractionsInitiateOidc.
 type PostIssuerProfilesProfileIDInteractionsInitiateOidcJSONBody = InitiateOIDC4VCRequest
 
 // PostIssueCredentialsJSONRequestBody defines body for PostIssueCredentials for application/json ContentType.
 type PostIssueCredentialsJSONRequestBody = PostIssueCredentialsJSONBody
+
+// PostCredentialsStatusJSONRequestBody defines body for PostCredentialsStatus for application/json ContentType.
+type PostCredentialsStatusJSONRequestBody = PostCredentialsStatusJSONBody
 
 // PostIssuerProfilesProfileIDInteractionsInitiateOidcJSONRequestBody defines body for PostIssuerProfilesProfileIDInteractionsInitiateOidc for application/json ContentType.
 type PostIssuerProfilesProfileIDInteractionsInitiateOidcJSONRequestBody = PostIssuerProfilesProfileIDInteractionsInitiateOidcJSONBody
@@ -99,6 +119,12 @@ type ServerInterface interface {
 	// Issue credential
 	// (POST /issuer/profiles/{profileID}/credentials/issue)
 	PostIssueCredentials(ctx echo.Context, profileID string) error
+	// Updates credential status.
+	// (POST /issuer/profiles/{profileID}/credentials/status)
+	PostCredentialsStatus(ctx echo.Context, profileID string) error
+	// Retrieves the credential status.
+	// (GET /issuer/profiles/{profileID}/credentials/status/{statusID})
+	GetCredentialsStatus(ctx echo.Context, profileID string, statusID string) error
 	// Initiate OIDC Credential Issuance
 	// (POST /issuer/profiles/{profileID}/interactions/initiate-oidc)
 	PostIssuerProfilesProfileIDInteractionsInitiateOidc(ctx echo.Context, profileID string) error
@@ -122,6 +148,46 @@ func (w *ServerInterfaceWrapper) PostIssueCredentials(ctx echo.Context) error {
 
 	// Invoke the callback with all the unmarshalled arguments
 	err = w.Handler.PostIssueCredentials(ctx, profileID)
+	return err
+}
+
+// PostCredentialsStatus converts echo context to params.
+func (w *ServerInterfaceWrapper) PostCredentialsStatus(ctx echo.Context) error {
+	var err error
+	// ------------- Path parameter "profileID" -------------
+	var profileID string
+
+	err = runtime.BindStyledParameterWithLocation("simple", false, "profileID", runtime.ParamLocationPath, ctx.Param("profileID"), &profileID)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter profileID: %s", err))
+	}
+
+	// Invoke the callback with all the unmarshalled arguments
+	err = w.Handler.PostCredentialsStatus(ctx, profileID)
+	return err
+}
+
+// GetCredentialsStatus converts echo context to params.
+func (w *ServerInterfaceWrapper) GetCredentialsStatus(ctx echo.Context) error {
+	var err error
+	// ------------- Path parameter "profileID" -------------
+	var profileID string
+
+	err = runtime.BindStyledParameterWithLocation("simple", false, "profileID", runtime.ParamLocationPath, ctx.Param("profileID"), &profileID)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter profileID: %s", err))
+	}
+
+	// ------------- Path parameter "statusID" -------------
+	var statusID string
+
+	err = runtime.BindStyledParameterWithLocation("simple", false, "statusID", runtime.ParamLocationPath, ctx.Param("statusID"), &statusID)
+	if err != nil {
+		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter statusID: %s", err))
+	}
+
+	// Invoke the callback with all the unmarshalled arguments
+	err = w.Handler.GetCredentialsStatus(ctx, profileID, statusID)
 	return err
 }
 
@@ -170,6 +236,8 @@ func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL 
 	}
 
 	router.POST(baseURL+"/issuer/profiles/:profileID/credentials/issue", wrapper.PostIssueCredentials)
+	router.POST(baseURL+"/issuer/profiles/:profileID/credentials/status", wrapper.PostCredentialsStatus)
+	router.GET(baseURL+"/issuer/profiles/:profileID/credentials/status/:statusID", wrapper.GetCredentialsStatus)
 	router.POST(baseURL+"/issuer/profiles/:profileID/interactions/initiate-oidc", wrapper.PostIssuerProfilesProfileIDInteractionsInitiateOidc)
 
 }
