+++ pkg/service/issuecredential/issuecredential_service.go
@@ -4,31 +4,27 @@ Copyright SecureKey Technologies Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-//go:generate mockgen -destination service_mocks_test.go -self_package mocks -package issuecredential -source=issuecredential_service.go -mock_names profileService=MockProfileService,kmsRegistry=MockKMSRegistry
+//go:generate mockgen -destination service_mocks_test.go -self_package mocks -package issuecredential -source=issuecredential_service.go -mock_names profileService=MockProfileService,kmsRegistry=MockKMSRegistry,vcStatusManager=MockVCStatusManager
 
 package issuecredential
 
 import (
 	"fmt"
-	"net/url"
 
 	"github.com/hyperledger/aries-framework-go/pkg/doc/verifiable"
 
 	"github.com/trustbloc/vcs/pkg/doc/vc"
 	"github.com/trustbloc/vcs/pkg/doc/vc/crypto"
-	cslstatus "github.com/trustbloc/vcs/pkg/doc/vc/status/csl"
 	"github.com/trustbloc/vcs/pkg/doc/vc/vcutil"
 	vcsverifiable "github.com/trustbloc/vcs/pkg/doc/verifiable"
 	vcskms "github.com/trustbloc/vcs/pkg/kms"
 	profileapi "github.com/trustbloc/vcs/pkg/profile"
-)
-
-const (
-	credentialStatus = "/credentials/status"
+	"github.com/trustbloc/vcs/pkg/service/credentialstatus"
 )
 
 type vcStatusManager interface {
 	CreateStatusID(vcSigner *vc.Signer, url string) (*verifiable.TypedID, error)
+	GetCredentialStatusURL(issuerProfileURL, issuerProfileID, statusID string) (string, error)
 }
 
 type vcCrypto interface {
@@ -85,9 +81,9 @@ func (s *Service) IssueCredential(credential *verifiable.Credential,
 		var status *verifiable.TypedID
 		var statusURL string
 
-		statusURL, err = url.JoinPath(profile.URL, profile.ID, credentialStatus)
+		statusURL, err = s.vcStatusManager.GetCredentialStatusURL(profile.URL, profile.ID, "")
 		if err != nil {
-			return nil, fmt.Errorf("failed to create sstatus URL: %w", err)
+			return nil, fmt.Errorf("failed to create status URL: %w", err)
 		}
 
 		status, err = s.vcStatusManager.CreateStatusID(signer, statusURL)
@@ -95,7 +91,7 @@ func (s *Service) IssueCredential(credential *verifiable.Credential,
 			return nil, fmt.Errorf("failed to add credential status: %w", err)
 		}
 
-		credential.Context = append(credential.Context, cslstatus.Context)
+		credential.Context = append(credential.Context, credentialstatus.Context)
 		credential.Status = status
 	}
