+++ pkg/service/verifycredential/revocation/testdata/university_degree.jsonld
