+++ pkg/service/verifycredential/testdata/revocation.jsonld
@@ -1,0 +1,16 @@
+{
+  "@context": [
+    "https://www.w3.org/2018/credentials/v1",
+    "https://w3id.org/vc/status-list/2021/v1"
+  ],
+  "id": "https://example.com/credentials/status/3",
+  "type": ["VerifiableCredential", "RevocationList2020Credential"],
+  "issuer": "did:trustblock:abc",
+  "issuanceDate": "2020-04-05T14:27:40Z",
+  "credentialSubject": {
+    "id": "https://example.com/status/3#list",
+    "type": "StatusList2021",
+    "statusPurpose": "revocation",
+    "encodedList": "H4sIAAAAAAAA_2IABAAA__-N7wLSAQAAAA"
+  }
+}
