+++ test/bdd/fixtures/profile/profiles.json
@@ -5,7 +5,7 @@
         "id": "i_myprofile_ud_P256k1",
         "name": "i_myprofile_ud_P256k1",
         "organizationID": "test_org",
-        "url": "https://test-issuer.com",
+        "url": "http://vc-rest-echo.trustbloc.local:8075",
         "active": true,
         "vcConfig": {
           "signingAlgorithm": "EcdsaSecp256k1Signature2019",
@@ -24,7 +24,7 @@
         "id": "i_myprofile_ud_es256_jwt",
         "name": "i_myprofile_ud_es256_jwt",
         "organizationID": "test_org",
-        "url": "https://test-issuer.com",
+        "url": "http://vc-rest-echo.trustbloc.local:8075",
         "active": true,
         "vcConfig": {
           "signingAlgorithm": "ES256",
@@ -43,7 +43,7 @@
         "id": "i_myprofile_ud_es384_jwt",
         "name": "i_myprofile_ud_es384_jwt",
         "organizationID": "test_org",
-        "url": "https://test-issuer.com",
+        "url": "http://vc-rest-echo.trustbloc.local:8075",
         "active": true,
         "vcConfig": {
           "signingAlgorithm": "ES384",
@@ -62,7 +62,7 @@
         "id": "i_myprofile_ud_es256k_jwt",
         "name": "i_myprofile_ud_es256k_jwt",
         "organizationID": "test_org",
-        "url": "https://test-issuer.com",
+        "url": "http://vc-rest-echo.trustbloc.local:8075",
         "active": true,
         "vcConfig": {
           "signingAlgorithm": "ES256K",
@@ -81,7 +81,7 @@
         "id": "i_myprofile_ud_p256",
         "name": "i_myprofile_ud_p256",
         "organizationID": "test_org",
-        "url": "https://test-issuer.com",
+        "url": "http://vc-rest-echo.trustbloc.local:8075",
         "active": true,
         "vcConfig": {
           "signingAlgorithm": "JsonWebSignature2020",
@@ -100,7 +100,7 @@
         "id": "i_myprofile_prc_P256k1",
         "name": "i_myprofile_prc_P256k1",
         "organizationID": "test_org",
-        "url": "https://test-issuer.com",
+        "url": "http://vc-rest-echo.trustbloc.local:8075",
         "active": true,
         "vcConfig": {
           "signingAlgorithm": "JsonWebSignature2020",
@@ -119,7 +119,7 @@
         "id": "i_myprofile_prc_p256",
         "name": "i_myprofile_prc_p256",
         "organizationID": "test_org",
-        "url": "https://test-issuer.com",
+        "url": "http://vc-rest-echo.trustbloc.local:8075",
         "active": true,
         "vcConfig": {
           "signingAlgorithm": "JsonWebSignature2020",
@@ -138,7 +138,7 @@
         "id": "i_myprofile_cp_p384",
         "name": "i_myprofile_cp_p384",
         "organizationID": "test_org",
-        "url": "https://test-issuer.com",
+        "url": "http://vc-rest-echo.trustbloc.local:8075",
         "active": true,
         "vcConfig": {
           "signingAlgorithm": "JsonWebSignature2020",
@@ -157,7 +157,7 @@
         "id": "i_myprofile_cp_p256",
         "name": "i_myprofile_cp_p256",
         "organizationID": "test_org",
-        "url": "https://test-issuer.com",
+        "url": "http://vc-rest-echo.trustbloc.local:8075",
         "active": true,
         "vcConfig": {
           "signingAlgorithm": "JsonWebSignature2020",
@@ -176,7 +176,7 @@
         "id": "i_myprofile_cmtr_p384",
         "name": "i_myprofile_cmtr_p384",
         "organizationID": "test_org",
-        "url": "https://test-issuer.com",
+        "url": "http://vc-rest-echo.trustbloc.local:8075",
         "active": true,
         "vcConfig": {
           "signingAlgorithm": "JsonWebSignature2020",
@@ -193,7 +193,7 @@
         "id": "i_myprofile_cmtr_p256",
         "name": "i_myprofile_cmtr_p256",
         "organizationID": "test_org",
-        "url": "https://test-issuer.com",
+        "url": "http://vc-rest-echo.trustbloc.local:8075",
         "active": true,
         "vcConfig": {
           "signingAlgorithm": "JsonWebSignature2020",
@@ -218,7 +218,7 @@
           "credential": {
             "format": ["ldp"],
             "proof": true,
-            "status": false
+            "status": true
           }
         },
         "oidcConfig": {
